+++ apps/dsa-web/src/pages/signal/index.ts
@@ -1,4 +1,5 @@
-export { default as SignalOverviewPage } from './SignalOverviewPage';
+export { default as SignalBriefingPage } from './SignalBriefingPage';
+export { default as SignalSettingsPage } from './SignalSettingsPage';
 export { default as QualityDashboard } from './QualityDashboard';
 export { default as ContentQueuePage } from './ContentQueuePage';
 export { default as AssetDetailPage } from './AssetDetailPage';